1.积分兑换奖品模块完成： * 积分奖品添加，商家可以做出更改编辑上下架权限 * 兑换奖品：如果用户积分足够，兑换奖品后奖品为未领取状态，兑换记录会出现一条奖品未兑换记录，在线下商家可以通过微信扫码进行兑换，其中兑换二维码使用Google的zxing生成兑换二维码，包裹微信回传用户信息的url，其中url因为太长微信可能会进行拦截因此使用短链转换，店家店员可以扫码执行兑换过程 2.商品购买获取积分模块完成： * 购买：用户登陆后每个商品详情页会展示一个商品二维码，用户在线下支付完成后，店员扫码就可以为该用户添加积分 3.新增错误状态码拦截 4.前端页面整合优化

+++ src/main/java/com/imooc/o2o/entity/Award.java
@@ -15,7 +15,7 @@ public class Award implements Serializable {
     /**
      * 积分奖品id，主键
      */
-    private Integer awardId;
+    private Long awardId;
 
     private String awardName;
 
@@ -28,7 +28,7 @@ public class Award implements Serializable {
      */
     private Integer point;
 
-    private Integer shopId;
+    private Long shopId;
 
     private Integer priority;
 
@@ -42,11 +42,11 @@ public class Award implements Serializable {
     private Integer enableStatus;
 
 
-    public Integer getAwardId() {
+    public Long getAwardId() {
         return awardId;
     }
 
-    public void setAwardId(Integer awardId) {
+    public void setAwardId(Long awardId) {
         this.awardId = awardId;
     }
 
@@ -82,11 +82,11 @@ public class Award implements Serializable {
         this.point = point;
     }
 
-    public Integer getShopId() {
+    public Long getShopId() {
         return shopId;
     }
 
-    public void setShopId(Integer shopId) {
+    public void setShopId(Long shopId) {
         this.shopId = shopId;
     }
